Fixed a bug that wouldn't let you edit cards before submitting the deck. Restructred component directories.
