+++ client/flip/src/compononents/common/CreateDeckButton/CreateDeckButton.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import styles from './CreateDeckButton.module.css';
+
+const CreateDeckButton = ({ onClick }) => {
+  return (
+    <div onClick={onClick} className={styles.createDeckButton}>
+      <div className={styles.header}>
+        Create Deck
+      </div>
+    </div>
+  );
+};
+
+export default CreateDeckButton;
