+++ client/flip/src/compononents/common/CreateDeckButton/index.jsx
@@ -1,14 +1,3 @@
-import React from 'react';
-import styles from './CreateDeckButton.module.css';
-
-const CreateDeckButton = ({ onClick }) => {
-  return (
-    <div onClick={onClick} className={styles.createDeckButton}>
-      <div className={styles.header}>
-        Create Deck
-      </div>
-    </div>
-  );
-};
+import CreateDeckButton from './CreateDeckButton';
 
 export default CreateDeckButton;
