+++ client/flip/src/compononents/common/DeleteDeckModal/DeleteDeckModal.module.css
@@ -1,4 +1,4 @@
-@import url('../../index.css');
+@import url('../../../index.css');
 
 .deleteDeckModal {
     height: 400px;
