+++ client/flip/src/compononents/common/ProfileLink/index.jsx
@@ -1,15 +1,3 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-import styles from './ProfileLink.module.css';
-
-const ProfileLink = () => {
-    return (
-        <li>
-            <Link className={`link ${styles.profileLink}`} to="/home">
-                Profile
-            </Link>
-        </li>
-    );
-};
+import ProfileLink from './ProfileLink';
 
 export default ProfileLink;
