+++ client/flip/src/compononents/common/SelectedDeck/SelectedDeck.module.css
@@ -1,4 +1,4 @@
-@import url('../../index.css');
+@import url('../../../index.css');
 
 .selectedDeckContainer {
     width: 80%;
