+++ client/flip/src/compononents/common/SelectedDeck/index.jsx
@@ -1,0 +1,3 @@
+import SelectedDeck from './SelectedDeck';
+
+export default SelectedDeck;
