+++ client/flip/src/compononents/common/SignupLoginToggle/index.jsx
@@ -1,14 +1,3 @@
-import React from 'react';
-import styles from './SignupLoginToggle.module.css';
-
-const SignupLoginToggle = ({ text, onClick }) => {
-    return (
-        <div className={styles.signupLoginToggle}>
-            <button className={styles.signupLoginToggleBtn} onClick={onClick}>
-                {text}
-            </button>
-        </div>
-    );
-};
+import SignupLoginToggle from './SignupLoginToggle';
 
 export default SignupLoginToggle;
