+++ client/flip/src/compononents/views/Browse/index.jsx
@@ -1,15 +1,3 @@
-import React from 'react';
-import BrowsePageContent from '../../common/BrowsePageContent';
-import Header from '../../common/Header/Header';
-import styles from './Browse.module.css';
-
-const Browse = () => {
-    return (
-        <div className={styles.browsePage}>
-            <Header />
-            <BrowsePageContent />
-        </div>
-    );
-};
+import Browse from './Browse';
 
 export default Browse;
