+++ client/flip/src/compononents/views/Study/index.jsx
@@ -1,15 +1,3 @@
-import React from 'react';
-import Header from '../../common/Header/Header';
-import StudyDeck from '../../common/StudyDeck';
-import styles from './Study.module.css';
-
-const Study = () => {
-    return (
-        <div className={styles.createDeckPage}>
-            <Header />
-            <StudyDeck />
-        </div>
-    );
-};
+import Study from './Study';
 
 export default Study;
